Refactor ReadBufferFromHDFS
